refactor: organize preview server modules
