+++ packages/cli/src/preview-server/constants.ts
@@ -1,0 +1,2 @@
+export const previewBuildCacheMarker = ".webstudio-preview-build";
+export const previewProcessOwnerFile = "preview-process.json";
