+++ packages/cli/src/preview-server/types.ts
@@ -1,0 +1,17 @@
+import type { ChildProcess } from "node:child_process";
+import type { ProjectPreviewMode } from "@webstudio-is/project-build/preview";
+
+export type PreviewMode = ProjectPreviewMode;
+
+export type PreviewServerOptions = {
+  host: string;
+  port: number;
+  mode?: PreviewMode;
+  cwd?: string;
+  imageDomains?: string[];
+};
+
+export type PreviewServerResult = {
+  url: string;
+  process: ChildProcess;
+};
